post-build: shuffle files around only when systemd is enabled

+++ buildroot/board/develer/develboard/post-build.sh
@@ -12,24 +12,30 @@ echo "Prepend AT91 nand flash header to bootstrap"
 	"$OUTPUTDIR"/sama5d4_xplained-nandflashboot-uboot-3.7.1.bin \
 	"$OUTPUTDIR"/at91bootstrap.bin
 
-echo "Remove sysvinit compatibility scripts"
-rm -rf "$ROOTFSDIR/etc/init.d"
-rm -rf "$ROOTFSDIR/etc/network"
+#
+# systemd specific
+#
 
-echo "Disable dropbear.service"
-rm -f "$ROOTFSDIR/etc/systemd/system/multi-user.target.wants/dropbear.service"
+if grep -q 'BR2_INIT_SYSTEMD=y' "${BR2_CONFIG}"; then
+    echo "Remove sysvinit compatibility scripts"
+    rm -rf "$ROOTFSDIR/etc/init.d"
+    rm -rf "$ROOTFSDIR/etc/network"
 
-echo "Enable dropbear.socket (socket activation)"
-ln -sf "$ROOTFSDIR/etc/systemd/system/dropbear.socket" "$ROOTFSDIR/etc/systemd/system/multi-user.target.wants/dropbear.socket"
+    echo "Disable dropbear.service"
+    rm -f "$ROOTFSDIR/etc/systemd/system/multi-user.target.wants/dropbear.service"
 
-echo "Enforce dynamic hostname regeneration through genhostname.service"
-rm -f "$ROOTFSDIR/etc/hostname"
-ln -sf "$ROOTFSDIR/etc/systemd/system/genhostname.service" "$ROOTFSDIR/etc/systemd/system/multi-user.target.wants/genhostname.service"
+    echo "Enable dropbear.socket (socket activation)"
+    ln -sf "$ROOTFSDIR/etc/systemd/system/dropbear.socket" "$ROOTFSDIR/etc/systemd/system/multi-user.target.wants/dropbear.socket"
 
-echo "Disable unnecessary services"
-find "$ROOTFSDIR" -type f -iname 'systemd-logind.service' -delete
-find "$ROOTFSDIR" -type f -iname 'systemd-remount-fs.service' -delete
-find "$ROOTFSDIR" -type f -iname 'systemd-udev-hwdb-update.service' -delete
-find "$ROOTFSDIR" -type f -iname 'systemd-user-sessions.service' -delete
+    echo "Enforce dynamic hostname regeneration through genhostname.service"
+    rm -f "$ROOTFSDIR/etc/hostname"
+    ln -sf "$ROOTFSDIR/etc/systemd/system/genhostname.service" "$ROOTFSDIR/etc/systemd/system/multi-user.target.wants/genhostname.service"
+
+    echo "Disable unnecessary services"
+    find "$ROOTFSDIR" -type f -iname 'systemd-logind.service' -delete
+    find "$ROOTFSDIR" -type f -iname 'systemd-remount-fs.service' -delete
+    find "$ROOTFSDIR" -type f -iname 'systemd-udev-hwdb-update.service' -delete
+    find "$ROOTFSDIR" -type f -iname 'systemd-user-sessions.service' -delete
+fi
 
 echo "Done"
